+++ packages/skeleton-common/src/classes/accordion.ts
@@ -1,9 +1,9 @@
 import { defineSkeletonClasses } from '../internal/define-skeleton-classes' with { type: 'macro' };
 
 export const classesAccordion = defineSkeletonClasses({
-	root: 'flex flex-col gap-2 w-full',
+	root: ['flex gap-2 w-full', 'data-[orientation=vertical]:flex-col data-[orientation=horizontal]:flex-row'],
 	content: 'py-2 px-4',
 	trigger: 'px-4 py-2 rounded-base w-full text-start hover:preset-tonal-primary',
-	item: 'flex flex-col gap-2',
+	item: ['flex gap-2', 'data-[orientation=vertical]:flex-col data-[orientation=horizontal]:flex-row'],
 	indicator: '',
 });